move form processing stuff out of __init__.py
